I accidentally  started the code with doubles instead of integers. I'm switching from doubles to integers instead of integers to doubles to show that I understand the concept.

+++ Calculator/Calculator.java
@@ -6,9 +6,9 @@ public class Calculator {
 
         Scanner scan = new Scanner(System.in);
         System.out.print("Enter a whole number: ");
-        double firstNum = scan.nextDouble();
+        int firstNum = scan.nextInt();
         System.out.print("Enter a second number: ");
-        double secondNum = scan.nextDouble();
+        int secondNum = scan.nextInt();
 
         System.out.println("\nEnter the symbol for the operation you would like to perform: \n+ (add), - (minus), * (multiply), / (divide)\n");
         String operation = scan.next();
